Moved helper function to under lib
